// src/app/Models/Product.ts
// export class Product{
//     pId: number;
//     pName: string;
//     pPrice: number;
//     pDiscount?: number;
//     pColors: string[];
//     pImage: string;
//     pInStock: boolean;
// }
export class Product {
    id: number;
    name: string;
    description: string;
    brand: string;
    gender: string;
    category: string;
    size: number[];
    color: string[];
    price: number;
    discount: number;
    inStock: boolean;
    items_left: number;
    imageURL: string;
    slug: string;
        
}